backend/requirements.txt
asgiref==3.11.0
charset-normalizer==3.4.4
contourpy==1.3.2
cycler==0.12.1
Django==5.2.10
djangorestframework==3.16.1
fonttools==4.61.1
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.2.6
packaging @ file:///C:/miniconda3/conda-bld/packaging_1761049101700/work
pandas==2.3.3
pillow==12.1.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
reportlab==4.4.9
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
